+++ AutoUsing/LibraryScanner.cs
@@ -1,16 +1,24 @@
+using System.IO;
+using System.Collections.Generic;
 namespace AutoUsing
 {
     public class LibraryScanner
     {
-        private LibraryScanner(){}
+        private LibraryScanner(string projectName, string projectDir){
+            this.projectName = projectName;
+            this.projectDir = projectDir;
+        }
 
         private string projectName;
         private string projectDir;
 
         
 
-        // public static List<LibraryScanner> GetScannersForAllProjects(){
-        //     foreach(var file in )
+        // public static List<LibraryScanner> GetScannersForAllProjects(string workspaceDir){
+        //     // var scanners = Directory.GetFiles(workspaceDir,"*.csproj",SearchOption.AllDirectories).
+        //     // foreach(var file in ){
+
+        //     // }
         // }