+++ src/Type.ts
@@ -1,0 +1,4 @@
+interface Type{
+    class : string;
+    namespace? : string;
+}