+++ src/test/playground/ShouldShowExtendAfterParentheses.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+namespace assets
+{
+    public class ShouldShowExtendAfterParentheses
+    {
+        void x(){
+            IEnumerable<string> x = null;
+            (((x))).Equals(2);
+        }
+    }
+}