+++ src/test/playground/ShouldShowExtensionsForSpaces.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+namespace assets
+{
+    public class ShouldShowExtensionsForSpaces
+    {
+        void x(){
+            IEnumerable<string> x = null;
+            x . Equals("123");
+        }
+    }
+}