+++ src/util.ts
@@ -1,3 +1,9 @@
 export function flatten<T>(arr : T[][]) : T[]{
     return arr.reduce((acc,val) => acc.concat(val),[]);
 }
+
+const debugging = false;
+
+export function AUDebug(str : string){
+    if(debugging) console.log(str);
+}